Restructure program & Add packaging system (#6)
