+++ .github/workflows/ci.yaml
@@ -23,16 +23,22 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
 
+      - uses: actions/cache@v3
+        id: cache
+        with:
+          path: ${{ env.pythonLocation }}
+          key: ${{ runner.os }}-python-${{ env.pythonLocation }}-${{ hashFiles('pyproject.toml') }}-test-v01
+
       - name: Install dependencies
+        if: steps.cache.outputs.cache-hit != 'true'
         run: |
           python -m pip install -U pip
-          pip install -r requirements.txt
+          pip install -e .
 
       - name: Run examples in Python-${{ matrix.python-version }}
-        working-directory: ./examples
         run: |
-          python3 example01.py
-          python3 example02.py
+          python3 examples/example01.py
+          python3 examples/example02.py
 
       - uses: actions/upload-artifact@v3
         with:
