+++ pycss_lem/__init__.py
@@ -1,0 +1,10 @@
+"""
+Open-source program written in Python3 for 2D slope stability analysis of
+circular surfaces by the limit equilibrium method (Fellenius and Bishop methods).
+"""
+
+__version__ = "0.1.0"
+
+
+from .automaticslipcircles import *
+from .onlyonecircle import *
