+++ pycss_lem/sliparcdiscretization.py
